make HeaderArrayInput component
